requirements.txt
ansible==2.9.2
ansible-runner==1.4.4
boto3==1.11.0
botocore==1.14.0
configparser==4.0.2
path==13.1.0
python-terraform==0.10.1
pywinrm==0.4.1
PyYAML==5.3.1
tabulate==0.8.6
